+++ src/featureforest/models/DinoV2/model.py
@@ -1,14 +1,10 @@
-from typing import Tuple
-
 import torch
 
 # from featureforest.utils.downloader import download_model
 from .adapter import DinoV2Adapter
 
 
-def get_model(
-    img_height: float, img_width: float, *args, **kwargs
-) -> DinoV2Adapter:
+def get_model(img_height: int, img_width: int, *args, **kwargs) -> DinoV2Adapter:
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     print(f"running on {device}")
     # get the pretrained model
@@ -18,8 +14,6 @@ def get_model(
     model.eval()
 
     # create the model adapter
-    dino_model_adapter = DinoV2Adapter(
-        model, img_height, img_width, device
-    )
+    dino_model_adapter = DinoV2Adapter(model, img_height, img_width, device)
 
     return dino_model_adapter
